-1.0.7b - added config options for progress panel time stamps - added subscriber for RL config changes - restructured how buy/sell and in progress panels are created and updated - added enum file for timestamp formats - updated tax calculation to use new tax rate  - buys/sells made and completed before update day will use old tax rate - removed unused tax calculation method in Transaction file - maybe some bug fixes

+++ src/main/java/com/flipper2/helpers/Numbers.java
@@ -5,7 +5,7 @@ import java.text.NumberFormat;
 
 public class Numbers
 {
-	public static String numberWithCommas(int number)
+	public static String numberWithCommas(long number)
 	{
 		return NumberFormat.getIntegerInstance().format(number);
 	}
@@ -29,7 +29,7 @@ public class Numbers
 	 * @param number The number to shorten.
 	 * @return The shortened number string.
 	 */
-	public static String toShortNumber(int number)
+	public static String toShortNumber(long number)
 	{
 		if (number > -100000 && number < 100000)
 		{
@@ -62,9 +62,9 @@ public class Numbers
 
 		shortNumber = Math.round(shortNumber * 10.0) / 10.0;
 
-		if (shortNumber == (int) shortNumber)
+		if (shortNumber == (long) shortNumber)
 		{
-			return String.format("%s%d%s", sign, (int) shortNumber, suffix);
+			return String.format("%s%d%s", sign, (long) shortNumber, suffix);
 		}
 		else
 		{
